requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dateutil==2.8.2
flask-restx==1.3.0
flasgger==0.9.7.1

